requirements.txt
flask
flask-cors
torch
transformers
wandb
uvicorn
python-dotenv
asyncpg
emoji